remove-date-picker - Update GeneralDogForm. Internalise the form schema. IO is now in terms of DogProfile.
