+++ src/app/user/(logged-in)/my-pets/add-dog/_components/add-dog-form.tsx
@@ -2,16 +2,12 @@
 
 import { BarkFormOption } from "@/components/bark/bark-form";
 import { DogProfile } from "@/lib/dog/dog-models";
-import { UTC_DATE_OPTION, parseDateTime } from "@/lib/utilities/bark-time";
 import { postDogProfile } from "../_actions/post-dog-profile";
 import { useRouter } from "next/navigation";
 import { RoutePath } from "@/lib/route-path";
 import { Err, Ok, Result } from "@/lib/utilities/result";
 import { CODE } from "@/lib/utilities/bark-code";
-import {
-  DogFormData,
-  GeneralDogForm,
-} from "../../_components/general-dog-form";
+import { GeneralDogForm } from "../../_components/general-dog-form";
 import { useToast } from "@/components/ui/use-toast";
 import { MINIMUM_TOAST_MILLIS } from "@/app/_lib/toast-delay";
 import { asyncSleep } from "@/lib/utilities/async-sleep";
@@ -22,20 +18,8 @@ export default function AddDogForm(props: { vetOptions: BarkFormOption[] }) {
   const { vetOptions } = props;
 
   async function handleValues(
-    values: DogFormData,
+    dogProfile: DogProfile,
   ): Promise<Result<true, string>> {
-    const {
-      dogBirthday: dogBirthdayString,
-      dogWeightKg: dogWeightKgString,
-      ...otherFields
-    } = values;
-    const dogBirthday = parseDateTime(dogBirthdayString, UTC_DATE_OPTION);
-    const dogWeightKg = parseFloat(dogWeightKgString);
-    const dogProfile: DogProfile = {
-      dogBirthday,
-      dogWeightKg,
-      ...otherFields,
-    };
     const { dogName } = dogProfile;
     toast({
       title: "Adding...",
@@ -68,7 +52,6 @@ export default function AddDogForm(props: { vetOptions: BarkFormOption[] }) {
     router.push(RoutePath.USER_MY_PETS);
   }
 
-  // WIP: This will break when we update GeneralDogForm
   return (
     <GeneralDogForm
       formTitle="Add Dog"
